2000-06-14  Andrew V. Samoilov  <sav@bcs.zp.ua>
	* panel.h, util.[ch] (struct WPanel, size_trunc, size_trunc_sep): size
	parameters and WPanel.total field became double, thus, total size
	of selected files/directories is shown in the mini status correctly
	when it exceeds 2 GB.

+++ src/panel.h
@@ -102,7 +102,7 @@ typedef struct {
     int      count;		/* Number of files in dir structure */
     int      marked;		/* Count of marked files */
     int      dirs_marked;	/* Count of marked directories */
-    long int total;		/* Bytes in marked files */
+    double   total;		/* Bytes in marked files */
     int      top_file;		/* The file showed on the top of the panel */
     int      selected;		/* Index to the selected file */
     int      reverse;		/* Show listing in reverse? */
